Revert "Create change report policy and invite submitter"
