--- src/libs/API/parameters/ChangeReportPolicyAndInviteSubmitterParams.ts
@@ -1,9 +1,0 @@
-type ChangeReportPolicyParams = {
-    reportID: string;
-    policyID: string;
-    reportPreviewReportActionID: string;
-    changePolicyReportActionID: string;
-    policyExpenseChatReportID: string;
-    policyExpenseCreatedReportActionID: string;
-};
-export default ChangeReportPolicyParams;
