+++ src/libs/actions/Policy/Policy.ts
@@ -77,7 +77,6 @@ import * as PhoneNumber from '@libs/PhoneNumber';
 import * as PolicyUtils from '@libs/PolicyUtils';
 import {goBackWhenEnableFeature, navigateToExpensifyCardPage} from '@libs/PolicyUtils';
 import * as ReportUtils from '@libs/ReportUtils';
-import type {OptimisticChatReport} from '@libs/ReportUtils';
 import type {PolicySelector} from '@pages/home/sidebar/FloatingActionButtonAndPopover';
 import * as PaymentMethods from '@userActions/PaymentMethods';
 import * as PersistedRequests from '@userActions/PersistedRequests';
@@ -112,7 +111,6 @@ type ReportCreationData = Record<
     {
         reportID: string;
         reportActionID?: string;
-        report: OptimisticChatReport;
     }
 >;
 
@@ -1072,7 +1070,6 @@ function createPolicyExpenseChats(policyID: string, invitedEmailsToAccountIDs: I
         // If the chat already exists, we don't want to create a new one - just make sure it's not archived
         if (oldChat) {
             workspaceMembersChats.reportCreationData[login] = {
-                report: oldChat,
                 reportID: oldChat.reportID,
             };
             workspaceMembersChats.onyxOptimisticData.push({
@@ -1122,7 +1119,6 @@ function createPolicyExpenseChats(policyID: string, invitedEmailsToAccountIDs: I
         const optimisticCreatedAction = ReportUtils.buildOptimisticCreatedReportAction(login);
 
         workspaceMembersChats.reportCreationData[login] = {
-            report: optimisticReport,
             reportID: optimisticReport.reportID,
             reportActionID: optimisticCreatedAction.reportActionID,
         };
